[IMP] partner_identification*: move _search_duplicate
