+++ partner_identification_gln/__manifest__.py
@@ -5,7 +5,7 @@
     "summary": """
         This addon extends "Partner Identification Numbers"
         to provide a number category for GLN registration""",
-    "version": "16.0.1.0.0",
+    "version": "16.0.1.0.1",
     "license": "AGPL-3",
     "author": "Acsone S.A.,Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/partner-contact",
